letter detail page 리팩토링
