+++ src/components/EmptyLetterDetail.jsx
@@ -1,0 +1,7 @@
+export default function EmptyLetterDetail() {
+  return (
+    <div>
+      <h2>You've taken the wrong path.</h2>
+    </div>
+  );
+}
